Move refresh token logic into base AuthStrategy
