+++ src/hakai_api/auth/base.py
@@ -7,6 +7,8 @@ from abc import ABC, abstractmethod
 from datetime import datetime
 from time import mktime
 
+import requests
+
 logger = logging.getLogger(__name__)
 
 
@@ -34,6 +36,16 @@ class AuthStrategy(ABC):
 
         Raises:
             ValueError: If credentials could not be obtained.
+        """
+        pass
+
+    @property
+    @abstractmethod
+    def client_type(self) -> str:
+        """Get the client type for this authentication strategy.
+
+        Returns:
+            The client type string (e.g., 'web', 'desktop').
         """
         pass
 
@@ -194,3 +206,46 @@ class AuthStrategy(ABC):
             return now > expires_at
         except (TypeError, ValueError):
             return True  # If we can't parse the expiry, consider it expired
+
+    def refresh_token(self, credentials: dict) -> dict | None:
+        """Refresh the access token using the refresh token.
+
+        Args:
+            credentials: Current credentials dictionary containing refresh_token.
+
+        Returns:
+            Updated credentials dictionary if successful, None otherwise.
+        """
+        if "refresh_token" not in credentials:
+            logger.debug("No refresh token available, cannot refresh")
+            return None
+
+        logger.debug("Attempting to refresh access token")
+
+        refresh_url = f"{self.api_root}/auth/refresh"
+        data = {
+            "refresh_token": credentials["refresh_token"],
+            "client_type": self.client_type,
+        }
+
+        try:
+            response = requests.post(refresh_url, json=data, timeout=10)
+
+            if response.status_code != 200:
+                logger.warning(f"Token refresh failed with status {response.status_code}")
+                return None
+
+            new_tokens = response.json()
+
+            # Update credentials
+            updated_credentials = credentials.copy()
+            updated_credentials["access_token"] = new_tokens["access_token"]
+            updated_credentials["expires_at"] = new_tokens["expires_at"]
+            updated_credentials["expires_in"] = new_tokens["expires_in"]
+
+            logger.info("Access token refreshed successfully")
+            return updated_credentials
+
+        except (requests.RequestException, json.JSONDecodeError, KeyError) as e:
+            logger.error(f"Token refresh failed with exception: {e}")
+            return None
